Make sure symbol_0 is selected by default
Change the initial class for symbol_0 (black) from "selected" to "selected-symbol-preview" so it's selected by default. Without this change there's a subtle bug where the user can appear to be making edits that aren't actually reflected in the underlying grid. 

Consider input 7f4411d.json. If the user copies the input grid and then starts clicking on the extra cells without selecting a color, the cells will change to black. When the user submits a grid that appears to be correct, they will be told their answer is incorrect because the changes weren't reflected in the underlying grid.